ref(backup): Split backup `helpers` module
This module has become something of a grab bag for various functions
used by the backup tooling. There are two classes of functionality that
we extract here:

1. Stuff related to encrypt/decrypt is moved into the new `crypto`
   module.
2. Stuff related to reflecting on the model graph is moved into the
   `dependencies` module, which is a more natural home for it.
