+++ src/sentry/api/endpoints/relocations/public_key.py
@@ -9,7 +9,7 @@ from sentry.api.api_owners import ApiOwner
 from sentry.api.api_publish_status import ApiPublishStatus
 from sentry.api.base import Endpoint, region_silo_endpoint
 from sentry.api.endpoints.relocations import ERR_FEATURE_DISABLED
-from sentry.backup.helpers import GCPKMSEncryptor, get_default_crypto_key_version
+from sentry.backup.crypto import GCPKMSEncryptor, get_default_crypto_key_version
 from sentry.utils.env import log_gcp_credentials_details
 
 logger = logging.getLogger(__name__)
