+++ src/sentry/backup/dependencies.py
@@ -642,3 +642,38 @@ def reversed_dependencies() -> list[Type[models.base.Model]]:
     sorted = list(sorted_dependencies())
     sorted.reverse()
     return sorted
+
+
+def get_final_derivations_of(model: Type) -> set[Type]:
+    """
+    A "final" derivation of the given `model` base class is any non-abstract class for the "sentry"
+    app with `BaseModel` as an ancestor. Top-level calls to this class should pass in `BaseModel` as
+    the argument.
+    """
+
+    out = set()
+    for sub in model.__subclasses__():
+        subs = sub.__subclasses__()
+        if subs:
+            out.update(get_final_derivations_of(sub))
+        if not sub._meta.abstract and sub._meta.db_table and sub._meta.app_label == "sentry":
+            out.add(sub)
+    return out
+
+
+# No arguments, so we lazily cache the result after the first calculation.
+@lru_cache(maxsize=1)
+def get_exportable_sentry_models() -> set[Type]:
+    """
+    Like `get_final_derivations_of`, except that it further filters the results to include only
+    `__relocation_scope__ != RelocationScope.Excluded`.
+    """
+
+    from sentry.db.models import BaseModel
+
+    return set(
+        filter(
+            lambda c: getattr(c, "__relocation_scope__") is not RelocationScope.Excluded,
+            get_final_derivations_of(BaseModel),
+        )
+    )
